requirements.txt
annotated-types==0.7.0
async-lru==2.0.4
bcrypt==4.2.1
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cssselect2==0.7.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
dnspython==2.7.0
evdev==1.9.1
filelock==3.17.0
fonttools==4.56.0
fuzzywuzzy==0.18.0
get-video-properties==0.1.1
gitdb==4.0.12
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
indexed_bzip2==1.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
Levenshtein==0.26.1
linkify-it-py==2.0.3
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
meson==1.7.0
meson-python==0.17.1
natsort==8.4.0
nltk==3.9.1
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
patchelf==0.17.2.1
pillow==11.1.0
pip-review==1.3.0
pipenv==2024.4.1
platformdirs==4.3.6
plumbum==1.9.0
proglog==0.1.10
psutil==7.0.0
pulsectl==24.12.0
py-gcode-metadata==0.2.0
pycairo==1.27.0
pyclip==0.7.0
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pyenchant==3.2.2
Pygments==2.19.1
PyGObject==3.50.0
PyGObject-stubs==2.12.0
pymongo==4.11.1
pyparsing==3.2.1
pyperclip==1.9.0
pyproject-metadata==0.9.0
Pyro5==5.15
pyspellchecker==0.8.2
python-dateutil==2.9.0.post0
python-wayland-extra==0.7.0
pyudev==0.24.3
pyusb==1.3.1
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
req2flatpak==0.2.0
requests==2.32.3
requirements-parser==0.11.0
rich==13.9.4
rpyc==6.0.1
serpent==1.41
setproctitle==1.3.5
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
speedtest-cli==2.1.3
streamcontroller-plugin-tools==2.0.2
streamcontroller-streamdeck==0.1.5
textual==2.1.2
tinycss2==1.4.0
tqdm==4.67.1
types-setuptools==75.8.0.20250225
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.3.0
usb-monitor==1.23
virtualenv==20.29.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
